+++ Nos/Views/RelayView.swift
@@ -7,47 +7,54 @@
 
 import SwiftUI
 import CoreData
+import Dependencies
 
 struct RelayView: View {
     @Environment(\.managedObjectContext) private var viewContext
     @EnvironmentObject private var relayService: RelayService
+    @ObservedObject var author: Author
     
     @State var newRelayAddress: String = ""
     
-    @FetchRequest(fetchRequest: Relay.allRelaysRequest(), animation: .default)
-    private var relays: FetchedResults<Relay>
-    
     @EnvironmentObject var router: Router
+    
+    @Dependency(\.analytics) private var analytics
     
     var body: some View {
         NavigationStack(path: $router.path) {
             List {
-                Section(Localized.relays.string) {
-                    ForEach(relays) { relay in
-                        Text(relay.address ?? Localized.error.string)
-                    }
-                    .onDelete { indexes in
-                        for index in indexes {
-                            let relay = relays[index]
-
-                            guard let address = relay.address else { continue }
-
-                            if let socket = relayService.socket(for: address) {
-                                for subId in relayService.activeSubscriptions {
-                                    relayService.sendClose(from: socket, subscription: subId)
-                                }
-
-                                relayService.close(socket: socket)
-                            }
-                            
-                            viewContext.delete(relay)
+                if let relays = author.relays?.allObjects as? [Relay] {
+                    Section(Localized.relays.string) {
+                        ForEach(relays) { relay in
+                            Text(relay.address ?? Localized.error.string)
                         }
-                        try! viewContext.save()
+                        .onDelete { indexes in
+                            for index in indexes {
+                                let relay = relays[index]
+                                
+                                guard let address = relay.address else { continue }
+                                
+                                if let socket = relayService.socket(for: address) {
+                                    for subId in relayService.activeSubscriptions {
+                                        relayService.sendClose(from: socket, subscription: subId)
+                                    }
+                                    
+                                    relayService.close(socket: socket)
+                                }
+                                
+                                analytics.removed(relay)
+                                author.remove(relay: relay)
+                                viewContext.delete(relay)
+                            }
+
+                            try! viewContext.save()
+                        }
+
+                        if author.relays?.count == 0 {
+                            Localized.noRelaysMessage.view
+                        }
                     }
-                    if relays.isEmpty {
-                        Localized.noRelaysMessage.view
-                    }
-                }            
+                }
                 Section(Localized.addRelay.string) {
                     TextField("wss://yourrelay.com", text: $newRelayAddress)
                         .autocorrectionDisabled()
@@ -60,7 +67,7 @@ struct RelayView: View {
                         CurrentUser.subscribe()
                     }
                 }
-                if relays.isEmpty {
+                if author.relays?.count == 0 {
                     Localized.noRelaysMessage.view
                 }
             }
@@ -93,8 +100,11 @@ struct RelayView: View {
             relay.createdAt = Date.now
             newRelayAddress = ""
 
+            CurrentUser.author.add(relay: relay)
+            
             do {
                 try viewContext.save()
+                analytics.added(relay)
             } catch {
                 // Replace this implementation with code to handle the error appropriately.
                 // fatalError() causes the application to generate a crash log and terminate. You should not
@@ -114,11 +124,11 @@ struct RelayView_Previews: PreviewProvider {
     
     static var previews: some View {
         NavigationStack {
-            RelayView()
+            RelayView(author: CurrentUser.author)
         }.environment(\.managedObjectContext, previewContext)
         
         NavigationStack {
-            RelayView()
+            RelayView(author: CurrentUser.author)
         }.environment(\.managedObjectContext, emptyContext)
     }
 }