+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Fix crash when going to the profile view, and then switching to the relays view
 
 ## [0.1 (6)] - 2023-03-08Z
 - Fixed follow / unfollow sync