+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Use only relays added in RelayView for sending and receiving events
 
+- Add PostHog analytics
+
 ## [0.1 (6)] - 2023-03-08Z
+
 - Fixed follow / unfollow sync
 - Reduced number of REQ sent to relays
 - Request profile metadata for users displayed on the Discover tab