+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Use only relays added in RelayView for sending and receiving events
+
+- Add PostHog analytics
+
+## [0.1 (6)] - 2023-03-08Z
+
 - Fixed follow / unfollow sync
 - Reduced number of REQ sent to relays
 - Request profile metadata for users displayed on the Discover tab
@@ -20,8 +26,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add the ellipsis button to NoteCard and allow the user to copy the NIP-19 note ID of a note.
 - Enabled button to copy user ID on Profile View
 - Fixed UI freezes when using many relays by moving event processing to a background thread.
+- Add a search bar to the discover tab that users can use to look up other users.
 - On relay remove, send CLOSE to all subs then disconnect and delete socket
 - Render user mentions in NoteCard
+- Replace the warning message to tell the user never to share their private key with anyone.
 
 ## [0.1 (5)] 2023-03-02